Fix errors caught by CI.
