+++ resolve/matrix/Csr.cpp
@@ -5,9 +5,8 @@
 
 #include "Csr.hpp"
 #include "Coo.hpp"
-#include <resolve/utilities/misc/IndexValuePair.hpp>
 #include <resolve/utilities/logger/Logger.hpp>
-#include <resolve/matrix/utilities.hpp>
+#include <resolve/matrix/Utilities.hpp>
 
 namespace ReSolve 
 {
