+++ resolve/matrix/Utilities.cpp
@@ -2,12 +2,47 @@
 #include <resolve/Common.hpp>
 #include <resolve/matrix/Coo.hpp>
 #include <resolve/matrix/Csr.hpp>
-#include <resolve/utilities/misc/IndexValuePair.hpp>
 #include <resolve/utilities/logger/Logger.hpp>
-#include "utilities.hpp"
+#include "Utilities.hpp"
 
 namespace ReSolve
 {
+  /// @brief Helper class for COO matrix sorting
+  class IndexValuePair
+  {
+    public:
+      IndexValuePair() : idx_(0), value_(0.0)
+      {}
+      ~IndexValuePair()
+      {}
+      void setIdx (index_type new_idx)
+      {
+        idx_ = new_idx;
+      }
+      void setValue (real_type new_value)
+      {
+        value_ = new_value;
+      }
+
+      index_type getIdx()
+      {
+        return idx_;
+      }
+      real_type getValue()
+      {
+        return value_;
+      }
+
+      bool operator < (const IndexValuePair& str) const
+      {
+        return (idx_ < str.idx_);
+      }  
+
+    private:
+      index_type idx_;
+      real_type value_;
+  };
+
   using out = io::Logger;
   namespace matrix
   {
